Start activity divided to fragments.
Need to fix bug with showing null_image when we swipe the last item from
the CardView. And fix null_image.
