--- app/src/main/res/drawable/gradient_header_background.xml
@@ -1,23 +1,0 @@
-<!--
-  Copyright 2014 Soichiro Kashima
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <gradient
-        android:angle="270"
-        android:endColor="#00616161"
-        android:startColor="#CC616161" />
-
-</shape>
