+++ app/src/main/res/layout/activity_start.xml
@@ -6,99 +6,42 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     tools:context="StartActivity">
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:weightSum="1"
     android:background="@color/grey">
-    <!--tools:context=".MyActivity">-->
-    <include layout="@layout/toolbar_actionbar" />
-    <Button
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="0 recipes available"
-        android:id="@+id/showAvailableRecipes"
-        android:layout_below="@+id/toolbar"
-        android:layout_centerHorizontal="true"
-        android:layout_marginLeft="110dp"
-        android:layout_marginTop="60dp"
-        android:visibility="invisible" />
+
+    <include layout="@layout/toolbar_actionbar"
+        android:id="@+id/include" />
+
+
+    <LinearLayout
+            android:orientation="vertical"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:layout_below="@+id/include"
+            android:layout_centerHorizontal="true"
+            android:id="@+id/startFragmentContainer">
+
+    </LinearLayout>
 
     <com.melnykov.fab.FloatingActionButton
         android:id="@+id/fab"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_gravity="bottom|right"
-        android:layout_margin="16dp"
-        android:src="@drawable/ic_add_white_24dp" />
+        android:layout_gravity="right"
+        android:src="@drawable/ic_add_white_24dp"
+        android:layout_alignParentBottom="true"
+        android:layout_alignParentRight="true"
+        android:layout_alignParentEnd="true"
+        android:layout_marginRight="23dp"
+        android:layout_marginBottom="33dp" />
 
-    <ImageButton
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/imageButton"
-        android:layout_gravity="center_horizontal|top"
-        android:layout_marginBottom="320dp"
-        android:layout_marginTop="56dp"
-        android:scaleType="centerCrop"
-        android:src="@drawable/interiorpng"
-        android:padding="0dp" />
-
-    <android.support.v7.widget.RecyclerView xmlns:android="http://schemas.android.com/apk/res/android"
-        android:id="@+id/recycler_view"
-        android:layout_width="match_parent"
-        android:layout_height="320dp"
-        android:layout_marginTop="110dp"
-        android:scrollbars="vertical"
-        android:layout_below="@+id/showAvailableRecipes"
-        android:layout_gravity="right|bottom">
-    </android.support.v7.widget.RecyclerView>
-
-    <TextView
-        android:layout_width="wrap_content"
-
-        android:layout_height="wrap_content"
-        android:textAppearance="?android:attr/textAppearanceLarge"
-        android:text="5"
-        android:id="@+id/bigNum"
-        android:textColor="@color/white"
-        
-        android:textSize="130dp"
-
-        android:layout_marginLeft="50dp"
-        android:layout_marginTop="60dp"
-        android:visibility="visible" />
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="рецептов доступно"
-        android:id="@+id/textView8"
-        android:layout_gravity="right|top"
-        android:layout_marginTop="170dp"
-        android:layout_marginRight="30dp"
-        android:textSize="23dp"
-        android:textColor="@color/white"
-        android:layout_alignRight="@+id/bigNum"
-       />
-
-    <RelativeLayout
-        android:layout_width="380dp"
-        android:layout_height="192dp"
-        android:layout_gravity="center_horizontal|top"
-        android:layout_marginTop="55dp"></RelativeLayout>
-
-    <ImageView
-        android:layout_width="222dp"
-        android:layout_height="359dp"
-        android:id="@+id/imageView"
-        android:layout_gravity="center"
-        android:src="@drawable/nulljpg"
-        android:layout_marginTop="90dp"
-        android:visibility="invisible" />
-</FrameLayout>
+</RelativeLayout>
     <!-- Nav drawer -->
     <include layout="@layout/navdrawer" />
 </android.support.v4.widget.DrawerLayout>
