+++ app/src/main/res/layout/fragment_empty_img_start.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:orientation="vertical" android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <ImageView
+        android:layout_width="222dp"
+        android:layout_height="359dp"
+        android:id="@+id/imageView"
+        android:src="@drawable/nulljpg"
+        android:visibility="visible"
+        android:layout_gravity="center"
+        android:layout_alignParentBottom="true"
+        android:layout_centerHorizontal="true"
+        android:layout_marginBottom="43dp" />
+
+
+</RelativeLayout>
